--- packages/utils/upgrade/resources/codemods/5.0.0/deprecate-helper-plugin-get-fetch-client.code.ts
@@ -1,21 +1,0 @@
-import type { Transform } from 'jscodeshift';
-import { changeImportSpecifier } from '../../utils/change-import';
-
-/**
- * change getFetchClient import from '@strapi/helper-plugin' to '@strapi/strapi/admin'
- */
-const transform: Transform = (file, api) => {
-  const { j } = api;
-
-  const root = j.withParser('tsx')(file.source);
-
-  changeImportSpecifier(root, j, {
-    oldMethodName: 'getFetchClient',
-    oldDependency: '@strapi/helper-plugin',
-    newDependency: '@strapi/strapi/admin',
-  });
-
-  return root.toSource();
-};
-
-export default transform;
